+++ src/app/app-routing.module.ts
@@ -1,11 +1,12 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { ExploreComponent } from './components/explore/explore.component';
-import { TweetComponent } from './components/tweet/tweet.component';
+import { CreateTweetComponent } from './pages/create-tweet/create-tweet.component';
+import { ExploreComponent } from './pages/explore/explore.component';
 
 const routes: Routes = [
 {path: '', component: ExploreComponent},
-{path: 'tweet', component: TweetComponent},
+{path: 'create-tweet', component: CreateTweetComponent},
+
 
 
 ];
