+++ src/app/app.component.html
@@ -1,11 +1,1 @@
-
-<h1>App Component</h1>
-
 <app-navbar></app-navbar>
-
-<app-explore></app-explore>
-
-<app-tweet *ngFor="let tweet of tweets" [tweet]="tweet"></app-tweet>
-
-
-
