--- src/app/components/explore/explore.component.html
@@ -1,2 +1,0 @@
-<p>we are in Exolor</p>
-
