--- src/app/components/explore/explore.component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-import { Tweet } from 'src/app/models/tweet';
-
-@Component({
-  selector: 'app-explore',
-  templateUrl: './explore.component.html',
-  styleUrls: ['./explore.component.scss']
-})
-export class ExploreComponent {
-  
-}
