+++ src/app/components/navbar/navbar.component.html
@@ -3,8 +3,8 @@
 <section>
   <nav>
     <ul class="menuItems">
-      <li><a [routerLink]="['/']" data-item='Explore'>Explore</a></li>
-      <li><a [routerLink]="['/tweet']" data-item='Tweet'>Tweet</a></li>
+      <li><a [routerLink]="['/']" data-item='Explore' routerLinkActive="true">Explore</a></li>
+      <li><a [routerLink]="['/create-tweet']" data-item='Create Tweet'>Create Tweet</a></li>
     </ul>
   </nav>
