+++ src/app/components/navbar/navbar.component.scss
@@ -25,9 +25,9 @@
   }
   
   nav {
-    margin: 25px;
+    margin: 20px;
     background: #f9f9f9;
-    padding: 16px;
+    padding: 8px;
 
     .menuItems {
       list-style: none;
