+++ src/app/components/tweet-item/tweet-item.component.html
@@ -1,4 +1,4 @@
-<h3>My Tweet</h3>
+<h3>we are in Tweet</h3>
 <section>
   <img [src]="tweet?.avatar" />
   <p>{{ tweet?.user }}</p>
