+++ src/app/components/tweet-item/tweet-item.component.ts
@@ -1,0 +1,13 @@
+import { Component, Input } from '@angular/core';
+import { Tweet } from 'src/app/models/tweet';
+
+@Component({
+  selector: 'app-tweet-item',
+  templateUrl: './tweet-item.component.html',
+  styleUrls: ['./tweet-item.component.scss']
+})
+export class TweetItemComponent {
+
+  @Input() tweet? : Tweet;
+
+}
