--- src/app/components/tweet/tweet.component.ts
@@ -1,13 +1,0 @@
-import { Component, Input } from '@angular/core';
-import { Tweet } from 'src/app/models/tweet';
-
-@Component({
-  selector: 'app-tweet',
-  templateUrl: './tweet.component.html',
-  styleUrls: ['./tweet.component.scss']
-})
-export class TweetComponent {
-
-  @Input() tweet? : Tweet;
-
-}
