+++ src/app/pages/create-tweet/create-tweet.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-create-tweet',
+  templateUrl: './create-tweet.component.html',
+  styleUrls: ['./create-tweet.component.scss']
+})
+export class CreateTweetComponent {
+
+}
