+++ src/app/pages/explore/explore.component.html
@@ -1,0 +1,3 @@
+<h3>we are in Explor</h3>
+
+<app-tweet-item *ngFor="let tweet of tweets" [tweet]="tweet"></app-tweet-item>
